friendlyelec: support huawei wifi 2 mini(E8372h)

+++ package/base-files/files/root/setup.sh
@@ -1,18 +1,27 @@
-#!/bin/ash
+#!/bin/sh
+
+logger "friendlyelec /root/setup.sh running"
+
+PLATFORM='sun8i|sun50i'
+if ! grep -E $PLATFORM /sys/class/sunxi_info/sys_info -q; then
+        logger "only support $PLATFORM. exiting..."
+        exit 0
+fi
 
 board=`grep "board_name" /sys/class/sunxi_info/sys_info`
 board=${board#*FriendlyElec }
-echo board=${board}
 
-#if [ "x${board}" = "xNanoPi-R1" ]; then
-echo "setup openwrt for ${board}..."
+logger "setup openwrt for ${board}..."
 cp -r /root/board/${board}/etc/* /etc/        
 uci commit             
 /etc/init.d/led restart
-/etc/init.d/network restart
-echo "done"
-#fi
 
-# setup ssid
-wifi_mac=`cat /sys/devices/platform/soc/1c10000.mmc/mmc_host/mmc2/mmc2:0001/mmc2:0001:1/ieee80211/phy0/macaddress`
-sed -i "s/OpenWrt/OpenWrt-${wifi_mac}/" /etc/config/wireless
+if grep ssid /etc/config/wireless -q; then
+        # setup ssid
+        path=`find /sys/ -name wlan* | tail -n 1`
+        wifi_mac=`cat ${path}/address`
+        sed -i "s/OpenWrt/OpenWrt-${wifi_mac}/" /etc/config/wireless
+fi
+
+/etc/init.d/network restart
+logger "done"
